backend/requirements.txt
torch
torchvision
python-dotenv
supabase